rename pkg/server to pkg/fuddle
